Wtf am i doing
