+++ main/website/static/main.js
@@ -1,0 +1,3 @@
+function test() {
+    window.location.href = "/stop"
+}
