+++ main/website/static/style.css
@@ -1,0 +1,9 @@
+#title {
+    text-align: center;
+    color:gray;
+}
+
+.Server-Life {
+    color: black;
+    text-align: left;
+}
