+++ server/ops.json
@@ -1,1 +1,8 @@
-[]
+[
+  {
+    "uuid": "0746958f-cc7c-441d-b9f2-f537a0c17505",
+    "name": "Bossotron1322",
+    "level": 4,
+    "bypassesPlayerLimit": false
+  }
+]
