+++ server/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Wed Jan 06 12:37:08 EST 2021
+#Fri Jan 08 15:26:54 EST 2021
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
