--- website/static/style.css
@@ -1,4 +1,0 @@
-#NameTitle {
-    text-align: center;
-    color:gray;
-}
