--- website/templates/button.html
@@ -1,8 +1,0 @@
-<html>
-    <header>
-        <title>Pyadmin Webpanel</title>
-        <h1 id = "NameTitle">Pyadmin Dev</h1>
-    </header>
-    <script src="{{url_for('static', filename='index.js')}}"></script>
-    <link rel= "stylesheet" type= "text/css" href="{{ url_for('static', filename='style.css') }}"/>
-</html>
